CH25 finish

+++ udemy/NodeJS/CH25/server/controller/feed.js
@@ -4,6 +4,7 @@ const path = require("path")
 const { body, validationResult } = require("express-validator")
 
 const Post = require("../models/post")
+const User = require("../models/user")
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1
@@ -18,13 +19,11 @@ exports.getPosts = (req, res, next) => {
         .limit(perPage)
     })
     .then((posts) => {
-      res
-        .status(200)
-        .json({
-          message: "Fetched posts",
-          posts: posts,
-          totalItems: totalItems,
-        })
+      res.status(200).json({
+        message: "Fetched posts",
+        posts: posts,
+        totalItems: totalItems,
+      })
     })
     .catch((err) => {
       if (err.statusCode) {
@@ -62,20 +61,31 @@ exports.createPost = (req, res, next) => {
 
   const title = req.body.title
   const content = req.body.content
+  let creator
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: "Maximilian",
-    },
+    creator: req.userId,
   })
   post
     .save()
     .then((result) => {
+      return User.findById(req.userId)
+    })
+    .then((user) => {
+      creator = user
+      user.posts.push(post)
+      return user.save()
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully",
-        post: result,
+        post: post,
+        creator: {
+          _id: creator._id,
+          name: creator.name,
+        },
       })
     })
     .catch((err) => {
@@ -131,6 +141,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404
         throw error
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized")
+        error.statusCode = 403
+        throw error
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl)
       }
@@ -160,8 +175,20 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404
         throw error
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized")
+        error.statusCode = 403
+        throw error
+      }
       clearImage(post.imageUrl)
       return Post.findByIdAndRemove(postId)
+    })
+    .then((result) => {
+      return User.findById(req.userId)
+    })
+    .then((user) => {
+      user.posts.pull(postId)
+      return user.save()
     })
     .then((result) => {
       res.status(200).json({ message: "Deleted" })
